Refactor: standardize admin component structure (folder per component, no index files)
