--- frontend/src/app/admin/_components/StatsCards/index.ts
@@ -1,1 +1,0 @@
-export { default } from './StatsCards';
