--- frontend/src/app/admin/_components/TrendCharts/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TrendCharts';
