--- frontend/src/app/admin/_components/index.ts
@@ -1,2 +1,0 @@
-export { default as StatsCards } from './StatsCards';
-export { default as TrendCharts } from './TrendCharts';
