+++ frontend/src/app/admin/events/_components/EventTable/EventTable.tsx
