+++ frontend/src/app/admin/settings/_components/CategoryItem/CategoryItem.tsx
