+++ frontend/src/app/admin/users/_components/UserDetailModal/UserDetailModal.tsx
